requirements.txt
colored_traceback==0.3.0
corner==2.2.2
easydict==1.10
EnergyFlow==1.3.2
h5py==3.7.0
ipdb==0.13.13
matplotlib==3.6.2
mplhep==0.3.26
numpy==1.23.5
pandas==1.5.2
prefetch_generator==1.0.3
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
rich==13.5.2
scikit_learn==1.2.0
scipy==1.11.2
tensorflow==2.9.0
torch==2.0.1
torch_ema==0.3
torchmetrics==1.0.3
torchvision==0.15.2
tqdm==4.64.1
wandb==0.13.6
